modify project

+++ backend/controllers/foodController.js
@@ -1,53 +1,54 @@
 import foodModel from "../models/foodModel.js";
-import fs from 'fs'
+import fs from "fs";
 
-// // add food item
-const addFood  = async (req,res) => {
-
+// add food item
+const addFood = async (req, res) => {
   let image_filename = `${req.file.filename}`;
 
   const food = new foodModel({
-    name:req.body.name,
-    description:req.body.description,
-    price:req.body.price,
-    category:req.body.category,
-    image:image_filename
-  })
-  try{
+    name: req.body.name,
+    description: req.body.description,
+    price: req.body.price,
+    category: req.body.category,
+    image: image_filename,
+  });
+  try {
     await food.save();
-    res.json({success:true,message:"Food added"})
-  }
-  catch(error){
-    console.log(error)
-    res.json({success:false,message:"Error occured"})
-  }
-}
-
-// // all food list
-const listFood = async(req,res) => {
-  try{
-    const foods = await foodModel.find({});
-    res.json({success:true,data:foods});
-  }
-  catch(error){
+    res.json({ success: true, message: "Food added" });
+  } catch (error) {
     console.log(error);
-    res.json({success:false,message:"error occured"});
+    res.json({ success: false, message: "Error occured" });
   }
-}
+};
 
-// // remove food item
-const removeFood = async(req,res) => {
+// all food list
+const listFood = async (req, res) => {
+  try {
+    const foods = await foodModel.find({});
+    res.json({ success: true, data: foods });
+  } catch (error) {
+    console.log(error);
+    res.json({ success: false, message: "error occured" });
+  }
+};
+
+// remove food item
+const removeFood = async (req, res) => {
   try {
     const food = await foodModel.findById(req.body.id);
-    fs.unlink(`uploads/${food.image}`,()=>{});
+    if (!food) {
+      return res.json({ success: false, message: "Food not found" });
+    }
+    fs.unlink(`uploads/${food.image}`, (err) => {
+      if (err) console.log(err);
+    });
 
     await foodModel.findByIdAndDelete(req.body.id);
-    res.json({success:true, message:"Food removed"});
+    res.json({ success: true, message: "Food removed" });
   } catch (error) {
-      console.log(error);
-      res.json({success:false,message:"error"});
+    console.log(error);
+    res.json({ success: false, message: "error" });
   }
-}
+};
 
-
-export {addFood, listFood, removeFood}
+export { addFood, listFood, removeFood };
